+++ api-tester/src/main/java/com/revenuecat/apitester/java/PurchaseOptionAPI.java
@@ -1,6 +1,5 @@
 package com.revenuecat.apitester.java;
 
-import com.revenuecat.purchases.google.GooglePurchaseOption;
 import com.revenuecat.purchases.models.PricingPhase;
 import com.revenuecat.purchases.models.PurchaseOption;
