+++ examples/purchase-tester/build.gradle
@@ -63,7 +63,6 @@ repositories {
 dependencies {
     implementation project(path: ':purchases')
     implementation project(path: ':feature:amazon')
-    implementation project(path: ':feature:google')
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
